layer-1-flattrade-python/requirements.txt
NorenRestApiPy
websocket-client
requests
pandas
kafka-python
python-dotenv
